requirements.txt
python_dotenv
requests
validators
requirements-parser
semantic-version
dohq-artifactory
pyyaml==5.1
SwarmManagement>=1.1.56
DockerBuildSystem>=1.1.39